+++ backend/flow/src/Flow/CallbackPayload.hs
@@ -32,21 +32,19 @@ instance ToJSON FlowCallbackEventV1Envelope where
   toJSON _ = unexpectedError "FlowCallbackEventV1Envelope: ToJson not implemented!"
   toEncoding FlowCallbackEventV1Envelope {..} =
     pairs
-      $  "version"
-      .= V1
-      <> "flow_instance_id"
-      .= flowInstanceId
-      <> "event_created"
-      .= eventCreated
+      $  ("version" .= V1)
+      <> ("flow_instance_id" .= flowInstanceId)
+      <> ("event_created" .= eventCreated)
       <> eventPairs event
     where
       eventPairs :: FlowCallbackEventV1 -> Series
       eventPairs Completed = "type" .= ("completed" :: Text)
       eventPairs Failed    = "type" .= ("failed" :: Text)
-      eventPairs (Rejected RejectedEvent { userName, message }) =
+      eventPairs (Rejected RejectedEvent { userName, mRejectMessage, mDocumentName }) =
         ("type" .= ("flow_rejected" :: Text))
           <> ("user_name" .= userName)
-          <> ("message" .= message)
+          <> ("message" .= mRejectMessage)
+          <> ("document_name" .= mDocumentName)
       eventPairs (AuthenticationAttempted AuthenticationAttemptedEvent {..}) =
         ("type" .= ("authentication_attempted" :: Text))
           <> ("user_name" .= userName)
@@ -58,17 +56,20 @@ instance FromJSON FlowCallbackEventV1Envelope where
   parseJSON = withObject "FlowCallbackEventV1Envelope" $ \o -> do
     event <- o .: "type" >>= getTypeSpecifics o
     FlowCallbackEventV1Envelope
-      <$> o
-      .:  "event_created"
-      <*> o
-      .:  "flow_instance_id"
+      <$> (o .: "event_created")
+      <*> (o .: "flow_instance_id")
       <*> pure event
     where
       getTypeSpecifics :: Object -> Text -> Parser FlowCallbackEventV1
       getTypeSpecifics _ "completed" = pure Completed
       getTypeSpecifics _ "failed"    = pure Failed
       getTypeSpecifics o "flow_rejected" =
-        Rejected <$> (RejectedEvent <$> o .: "user_name" <*> o .: "message")
+        Rejected
+          <$> (   RejectedEvent
+              <$> (o .: "user_name")
+              <*> (o .: "document_name")
+              <*> (o .: "message")
+              )
       getTypeSpecifics o "authenication_attempted" =
         AuthenticationAttempted
           <$> (   AuthenticationAttemptedEvent
@@ -88,7 +89,8 @@ data FlowCallbackEventV1
 
 data RejectedEvent = RejectedEvent
   { userName :: UserName
-  , message ::  Text
+  , mDocumentName :: Maybe DocumentName
+  , mRejectMessage ::  Maybe Text
   }
   deriving (Eq, Generic, Show)
 
@@ -128,4 +130,3 @@ instance ToJSON OnfidoProviderData where
 
 instance FromJSON OnfidoProviderData where
   parseJSON = genericParseJSON aesonOptions
-