+++ backend/migrate/inc/AppDBMigrations.hs
@@ -303,4 +303,5 @@ kontraMigrations =
   , migrateNullableDocumentNameInEvent
   , addProviderMethodToFlowEIDAuthentications
   , createTableFlowEidServiceTransactions
+  , addDetailsJsonColumnToFlowEvents
   ]